Call _child_changed after child has been set
"set_child" should call "_child_changed", which
is, well, the internal routine handling changes
in children nodes.
In addition to the current behavior this will
add the following improvements:
(1) State which was invalidated by the change in
the children node such as statistics
and output column references will now be
cleared and recomputed when required.
(2) "_child_changed" will also call
"_on_child_changed" which applies changes to the
node state specific to the type of node at hand.