Fix derive_statistics_from on predicate nodes
The column IDs should be retrieved from the passed-in
children and not from the node itself since

(1) the resulting column IDs may not exist or give
wrong results if passed-in-child and this-child
do not agree on a common source of information
(e.g., stored table node).

(2) retrieving the column-IDs may fail since
this-child might not be present in case this
predicate node was detached from the LQP.
This can occur as part of the predicate
reordering optimizer rule.