TPC-H Q1 date selection: filter on shipdate (#1152)
As of now in TPC-H Q1, we select all dates before-equals the max shipdate generated by the tpch datagen: `      WHERE l_shipdate <= '1998-12-01'`
This means we have a filter that does not filter out any rows: ...

The TPC-H specification uses this selection: `l_shipdate <= date '1998-12-01' - interval '[DELTA]' day (3)`
where DELTA is `is randomly selected within [60. 120].` (source: TPC-H spec 2.4.1.3)

The PR changes the selection value as if we would choose an interval of 90 days. This is similar to what HyPer did for their last VLDB paper.

The performance impact is neglectable since the filter returns now 5912 instead of 6004 rows on a scale factor of 0.001 (should scale linearly with larger scale factors). According to the specification, the 90 days interval should return 96% of the tuples.