[DYOD18/19] [constraints/master <--] Refactored Commit Phase (#1565)
This PR introduces our last change. As shown in our talk it is necessary to reorder the commands executed during the commit. In the new state, the checks for uniqueness are executed when the transaction manager fires the callback. This ensures that only one check at a time is executed. The test ConcurrentInsertInsertRace shows how a multithreaded scenario would look like, in which the old implementation would have failed.