[JIT] Add dictionary support for jittable table scans (#1532)
This PR introduces the support to compare values from dictionary-compressed segments via their value ids.

The difficult part when implementing this is that the same specialized function is used for all chunks. The encodings of the chunks can vary (e.g., the last chunk is unencoded) so that a specialized function cannot be always used for all chunks.
The encodings of the first chunk are taken as a reference for specialization. All chunks with different encodings cannot be processed with the specialized function. The fallback for these chunks is to interpret the operators.

The main changes to allow comparisons via value ids include:

    A new ValueID Jit data type
    Different JitReadTuples class members are not const as the result data type must be alterable.
    The JitAwareLQPTranslator checks for every expression whether a comparison via value ids could be possible
    The JitReadTuples operator incorporates the logic to check whether a potential value id expression can be used and updates the expression accordingly.
    A new JitReadTuples::before_specialization() method is necessary to update the JitExpressions before specialization.