Force pipeline users to deal with transaction aborts (#1631)
Fixes #1610. Other than originally thought, the pipeline already has some handling of failed transactions, but it is incomplete. With get_result_tables(), the caller was supposed to check failed_pipeline_statement() before continuing. The shortcut, get_result_table() did not work at all.

With this PR, the pipeline returns a pair of a pipeline status and the table(s). This forces the caller to deal with the possibility of a transaction abort (or at least actively ignore the bool).