[JIT] Remove necessity to access nullable information in JitAwareLQPTranslator (#1599)
This PR removes the necessity to access nullable information in the JitAwareLQPTranslator.

This requires that the nullable information in the JitTupleEntry instances, which are created during translation, must be updated before specialization. This is achieved by calling before_specialization() function on all JitOperators to allow them to update the nullable information in the JitExpression which they use.
The JitExpressions are updated recursively in update_nullable_information.