JoinIndex support index on both sides (#1727)
This PR integrates the possibility to use the JoinIndex with the mandatory index structures in the output table of the left input operator. Previously the index structures had to be on the right side.
Now, the JoinIndex constructor expects the specification of the join input operator which output table contains the index structures (const JoinInputSide index_side). By default, the index_side is JoinInputSide::Right.