Remove function Table::chunks() (#1730)
This PR has its root in #1686, which implements std::atomic_load and std::atomic_store for the vector _chunks in Table. Latter PR is in favor of the MvccDeletePlugin, which removes chunks from the _chunks vector.

Each read-/write-access on _chunks has to happen via atomics functions. People might forget about this and break thread-safety unwittingly.
Because of this, we remove the chunks() function from Table.  This limits all critical access to the _chunks vector to a single class (Table).