Hash Join: improve offset calculation (#1738)
As found by @mrks, the offset calculation of the hash join can be quite slow for very large inputs (e.g., TPC-H with a scale factor of 1000).

This PR changes the previous code that was rather randomly aggregating values and thus created a lot of cache misses. The new code iterates the input data twice, but fully sequential.

The change is not measurable for small joins.
For a SF=100 join and Q21, the offset part of the largest join went from 1.75s and 1.57s to now 0.29s and 0.28s. For larger inputs, the change should be more drastic.