[JIT] Fix bug #1741 - no support for COUNT(*) without group by (#1775)
This pr fixes the issue that Aggregate in JIT could not handle COUNT(*) without a group by column.

JitAggregate always requires a value as a parameter for an aggregate function. Since COUNT(*) does not have a parameter, we just passed in the first value of the tuple as a dummy. This caused a problem if the tuple has no values and is now solved by adding a constant value to the tuple and using it as a parameter.