Fix sqlite comparison (#1784)
By default, LIKE is case-insensitive in sqlite. This PR changes that behavior to match Hyrise's.

Also, it fixes the comparison of all-NULL columns.

There is still a mismatch in the table sizes between the imdb_data set that we are using and the one linked in the benchmark. That leads to some results being NULL as described in #1409. I don't see this as a big issue, as we are only using the JOB for internal comparisons.