TPC-DS Table Generator (#1750)
This PR adds an in-memory table generator to hyrise that conveniently creates tables filled with tpcds data.
The code relies heavily on the TableBuilder class and on the officially provided tpcds-tools. These tools are written in C and use quite a lot of global state and macros. Additionally there are some interesting design decisions, eg. letting the print function decide whether to print the given value or NULL. Therefore I had to write some code that's less pretty than I would like.