Don't increase commit id for read-only transactions (#1802)
During the work on the TPC-C, I realized that we should not commit transactions that did not change anything. There, my solution was to simply throw the transaction context away. However, it is much better to solve the underlying problem:

If a transaction does not have registered read-write operators, committing it should return immediately, leaving the global last commit ID untouched.