Fix performance regression in JOB (#1848)
fixes #1839 - we still have some queries that run slower, but the -99% parts are gone. Given that we don't really optimize for the JOB, that should be good enough for now.

Splitting up x LIKE '...' OR x LIKE '...' into two predicates (#1791) enables the TableScan to use the ColumnLikeTableScanImpl. That implementation first builds a bitmap of the entire dictionary, then checks the value id against that bitmap. If the number of input rows is significantly lower than the size of the dictionary (because the input has already been filtered), that is a bad idea. Now, we only scan the entire dictionary if the filtered segment has no more unique values in the dictionary than rows in the input.