Better estimation for correlated parameters (#1863)
Estimating correlated parameters is tricky. Example: SELECT c_custkey, (SELECT AVG(o_totalprice) FROM orders WHERE o_custkey = c_custkey) FROM customer. If the subquery was executed for each customer row, assuming that the predicate has a selectivity matching that of searching for a single value would be reasonable. However, it is likely that the SubqueryToJoinRule will rewrite this query so that the CorrelatedParameterExpression will turn into an LQPColumnExpression that is part of a join predicate. However, since the JoinOrderingRule is executed before the SubqueryToJoinRule, it would create a different join order if it assumes orders to be filtered down to very few values.

For now, return a high selectivity (90%) for predicates that contain correlated parameters.