Make left/right side of joins deterministic (#1865)
To make the input sides more deterministic, we make sure that the larger input is on the right side. This helps future rules to identify common subplans. For plans with equal cardinalities, this might still result in non-deterministic orders. Deal with it once it becomes a problem - maybe use the hash?

Also, adds the column names to the printed version of MockNodes. Trying to understand why two seemingly identical MockNodes were considered different took me longer than it should.

No performance difference in TPC-H yet - that will come with future rules.