Forward a chunk's ordered-by information to GetTable (#1866)
Currently, the order-by information is not carried through to the table scan. This PR adapts (i) the abstract table generator to add the sorted information to the chunks (in case sorting is requested for the benchmark) and (ii) the get table operator to forward and adapt (in case of column pruning) this information.

For the benchmarks, basically nothing changed. Only TPC-H Q3 has a qualifying scan, but the impact is not significant.