Update column statistics when chunks are pruned (#1859)
If a chunk is pruned, we update the table statistics. This is so that the selectivity of the predicate that was used for pruning can be correctly estimated. Example: For a table that has sorted values from 1 to 100 and a chunk size of 10, the predicate x > 90 has a selectivity of 10%. However, if the ChunkPruningRule removes nine chunks out of ten, the selectivity is now 100%. Updating the statistics is important so that the predicate ordering can properly order the predicates. For the column that the predicate pruned on, we remove chunk.size() values that do not match the predicate from the histogram, assuming equidistribution among the pruned values. The other columns are simply scaled to reflect the reduced table size.