Retire JIT (#1883)
This retires the JIT component. We make this step with a heavy heart as both students and ourselves have put a significant amount of time into it. The following reasons compel us to make this step:

    With the dependencies to LLVM and the modified LLVM source files, upgrading to a newer LLVM version is difficult. We are still stuck with LLVM 7, while LLVM 8 has been out since March and LLVM 9 since September.

    The performance of our JIT code is significantly worse than that of the non-JITted code (see below). With six months spent optimizing the JIT code, it became clear that the underlying architecture is a bottleneck. By now, even TPC-H query 1, which was the main show case, does not have an advantage anymore.

    We are lacking the resources to work on this and improve it to a point where it delivers the performance generally expected from a JIT engine.

The following features have been removed as they are no longer used:

    SQLPipelineBuilder::with_lqp_translator, as we only have one LQPTranslator left
    InsertLimitInExistsRule, as the Limit operator is generally bad for performance in our non-JITted execution

lqp_subplan_to_boolean_expression is now unused, but since LQP utilities are always good to have, I would vote for leaving it.