CI: Replace TPC-H test with --verify (#1885)
When we started implementing the TPC-H, our performance was too bad to even execute SF 1 as part of CI. Thus, we had different scale factors that were used for the tests. I suspect that the time spent building those different TPC-H databases now surpasses that of the actual queries. Let's see how long ./hyriseBenchmarkTPCH --verify takes on Jenkins.

As a bonus, we will now execute TPC-H on tsan runs with the scheduler, so that tsan actually makes some sense.

Also removes the tests for TPC-H/DS data generation from the sanitizer runs. As those won't change and are unlikely to suddenly develop asan/ubsan/tsan issues, it feels like we are wasting CI time there.