Upgrade to clang-9 (#1897)
* Upgrades to clang-9, so that OS X builds on the most current versions work again
* Brings in the two changes from #1881 that have the biggest effect on the linking time
* Another round of clang-tidy fixes
* Fixes a couple of threading issues, found by tsan

clang-tidy and tsan have really improved a lot in this release. I am still surprised at the complexity of problems now identified.

This is more hacky than I want it to be, mostly because boost is broken in two ways that are only identified by recent compilers. Unfortunately, Ubuntu is shipping with antique boost versions, so we need to do some backporting.