Fix LQP outputs (#1907)
With both visit_lqp_upwards and the new ColumnPruningRule, we depend on the output links of AbstractLQPNodes to be correct. We now make sure that the optimizer has exclusive ownership so that no remnants of the old plan remain.

This also adds a verify_lqp method to the optimizer (which we should have had a long time ago). While it is not perfect, it should help us catch issues as soon as a rule breaks something, not just once a following rule or the LQPTranslator stumbles upon the issue.