Modifications to LZ4 iterable (#1906)
When not erasing LZ4 iterables, it turned out that certain std algorithms cannot use the iterators. With @mrks 's help, these issues have been fixed. This change further allows to use unerased LZ4 segments, which are of relevance for my compression selection work and @krichly's spatial work.

Further, it adds a slight change to the caching of blocks. With this change, an iterable that is accessed multiple times, can still profit from LZ4-block caching (will be relevant for upcoming changes to the segment accessors).