Refactor ColumnPruningRule, add basic constraints (#1854)
There are some cases that the current ColumnPruningRule does not catch. Most importantly, if a column stops being useful at a certain point in the LQP, it is not pruned at the next opportunity.

This new rule:
* Performs a top-down search for actually required expressions and erases expressions from ProjectionNodes if they are unnecessarily forwarded
* Continues to prune columns from StoredTableNodes
* Identifies inner joins that are actually semi joins (because all columns from one side remain unused). This enables future optimization.

For the last point, we need information about constraints (UNIQUE, PRIMARY). I have partially integrated the constraint definitions from #1570; however, these constraints are for optimization only and are not yet enforced.