Retrieve access frequency of cached elements (#1932)
This PR introduces the frequency() method to the cache. The GDFS cache (which is the only one used right now) has a nice F in the name and we should use that.

The function is not implemented for other cache. On the one hand, they aren't used anyways. On the other hand, as far as I can tell they do not store any frequency information.