Optimizer rule: remove dependent group-by columns (#1929)
This PR partially implements the choke point 1.4 of Boncz et al.'s "TPC-H Analyzed: Hidden Messages and Lessons Learned from an Influential Benchmark".

The concept is simple. Whenever we have the primary key(s) and other columns (of the same table) in the group by list, we can remove the other columns are they are functionally dependent and thus do not have an effect on the grouping.
Often, columns are grouped to allow later access to them (see Q10). To cover this case, we added the ANY() aggregation function which is a sole internal function that returns "any" element of the group (it is thus only used when we know there is only a single distinct value).

We only "partially" implemented the rule, as Boncz et al. also list potential optimization for which we require foreign keys. Due to the current lack of these, we can only optimize TPC-H Q10 and Q18.
For Q18, the runtime of the aggregate only contributes a very small part of the runtime. Thus, there are no big gains to expect.