Fix crashing randomized test: missing fixtures keeps storage manager alive (#1982)
#1968 changed a test to be non-fixtured (changed from TEST_F to TEST) which uses the storage manager. This was dump because the storage manager is not reset after the test.
For the randomized tests of the CI, by chance there could be tests following that also create a table int_int (e.g., the union tests). This will crash since the table already exists.

I grepped for tests that contain the storage_manager and do not use fixtures. The found tests look good. I also looked for cases that do not inherit from BaseTest, but did not find one that uses the storage manager.