Remove dictionary() from FixedStringDictionarySegment (#1973)
In the discussion of #1968 we found a rather strange method named dictionary() within the fixed string vector. While this method has a misleading name, it also copied the whole vector into a new vector. This has (probably) been done since the other dictionary segment implementation did the same.

This PR removes this method altogether. The like matcher needed to be adapted to handle both pmr_string as well as std::string_view.