[DYOD] Add null value specification to SyntheticTableGenerator (#1987)
This will add a way to specify which part (ratio) of the generated table will be filled with null values. Since this would add another specification on per-column-basis to generateTable we decided to factor out all specifications that apply per column into one vector of a newly introduced struct ColumnSpecification.