TPC-H: Don't sort binary tables again, use cache by default (#2023)
The binary dumps have no information if the entire table is sorted. As such, we currently re-sort and re-encode TPC-H tables. By checking if the table is already sorted, we can avoid this step and reduce the benchmark startup time from 0m30.447s to 0m16.147s. This is not the perfect solution and storing the flag in the table would be nicer.