Improve advance() for SimdBp128 iterators (#2015)
@krichly found that sorted searches (on an experimental branch that does not have the "being an insert iterator" issue, cf. #1531) on segments that use SimdBp128 vector compression are unexpected slow. The reason was that the underlying access to the vector via the provided iterators does not support random access.

This PR changes the current advance() implementation to no longer do n single steps but rather be an actual advance() operation. Before that, the iterator was rather complex as it did caching and extracted the data by itself. Now, we pass the compressor to the iterator and use that. The decompressor has already code for random jumps and does caching as well.

Further change is that decompressors are passed by value to the iterators (no longer shared_ptr to a shared decompressor) as they internally cache blocks. This required some added constructors and assignment operators.

Please note, that #1531 still applies and the changes to the advance() method will not be observable with the current master.