AggregateHash: Short group-by string optimization (#2010)
During the group-by phase, the hash-based aggregate has to generate a unique key for each result group. For a single group-by column, this unique key could be the column type. However, as we have to support multiple group-by columns that could have different types, we use uint64_ts (aka. AggregateKeyEntry) instead. By default, these are generated similar to the dictionary encoding, i.e., by building a map of all values seen in the input. Of course, looking up the existing "value id" in that map (or even worse, inserting a new one) is expensive. Since #1957, we have an optimization for int32_ts, which can be converted into uint64_ts.

With this PR, we add a similar optimization for strings shorter than five characters. For those, there are 4'311'810'305 different combinations. We can directly convert these strings into IDs without using a map and the remaining "value id" space is still sufficiently large.

I'll be the first to admit that this is a specific optimization and doesn't make the code any nicer. However, unless we come up with a new, smarter way for grouping columns, we will need to do things like this to somehow improve the pathetic performance of the hash-based aggregate.