Implement advance() for RunLength Segment Iterables (#2038)
Currently, the RunLength segment iterables do not support a proper advance() but rather iteratively call ++ or --.

This PR implements a "proper" advance() method. Further, the search for end positions has changed. Previously, the "next" chunk offset was resolved via a linear search, no matter how large the step from the previous to the current offset was (it worked well for most cases, but for large jumps a binary search is much faster, especially for SF 10).
Now, we first estimate a threshold of when to use linear/binary search and then use this threshold to determine which search method shall be used.