Catch duplicate aggregate functions (#2057)
These queries are currently broken because the SQLTranslator incorrectly identifies the aggregate function as having already been calculated (see #1902):

SELECT COUNT(*) FROM (SELECT COUNT(*) FROM t WHERE a <= 1234) t2
SELECT COUNT(a) FROM (SELECT a, COUNT(a) FROM t GROUP BY a) t2
SELECT AVG(a) FROM (SELECT a, AVG(a) FROM t GROUP BY a) t3

As long as this (1) does not produce wrong results and (2) does not get in the way of experiments or benchmarks, it feels more like a cosmetic issue. This PR introduces a check for these cases.