Remove make_(shared|unique)_by_data_type (#2066)
In most places, we now use resolve_type to change a run-time data type into a compile-time type. resolve_type is more flexible and does not force people to create unintuitive Impl classes.

This PR removes the remaining uses of make_(shared|unique)_by_data_type in order to reduce the amounts of meta magic in the code basis.