Improve advance() for FrameOfReference Iterable (#2062)
This PR implements a better version of advance() for FrameOfReference iterables and furthermore improves the non-pos-list iterable.

The first implementation of advance() revealed that we update too many temporary values which was actually not necessary. Now, the non-pos-list iterator uses (similar to the pos-list iterator of the current master) pointers which avoids several iterator copies.

I know this iterable contains pointers, which we'd like to avoid. But I don't see a real alternative here and I don't think we should ignore the performance advantages.

Since we still don't use advance() a lot (#1531), the performance difference is mainly caused by switching to pointers.