Fix parallel insertion of NULLable values, use proper NULL in TPC-C (#2051)
In a ValueSegment, NULLs are stored in a pmr_vector<bool>. Writing to values within the same block (i.e., where n / 8 is equal) is not thread-safe. For now, I added a mutex that is acquired when the NULL values are modified. I'm not really happy with the solution, but the alternative would be to exchange the data structure. Not only here, but also in every class that creates a ValueSegment.

This came up when I added NULL support for TPC-C (instead of using -1). As such, this PR fixes #1811.