Expression: faster equals and hash (#2068)
For operator== on an expression, it makes sense to first compare the expression itself before recursing into the inputs.

When hashing, we do not need to consider each and every part of the expression tree. As a hash comparison is always followed by a call to operator==, we will have to recurse anyway and while operator== can use early outs once it detected a difference, hash cannot.