Better information during table generation (#2084)
I am working on compression schemes and was surprised that (the apparently not necessary) re-encoding (importing LZ4 as binary and using LZ4 as the compression for all columns) took quite a long time (can be easily minutes for large data sets).

Everything worked as expected, but our encoding step also generates the pruning statistics. This PR simply adds that information to the output as it might be confusing to users (as it was to me).