NULLs as std::optional in FoR segments (#2070)
This PR stores the bool vector for NULL values as an std::optional<pmr_vector<bool>>. The performance slightly degrades for GCC (I'd say it's noise because I've seen slight different numbers before) and improves for Clang. I don't think this improvement is too surprising, because the branch should be well predictable and even jumps in a small bool vector don't come for free.
But I am still suprised that the improvement is so "large" for Clang.

The storage overhead is strongly depending on the actual data (unsurprisingly). For TPC-H with all integers as FoR-encoded and FSBA we save a little less than 5%, for SIMDBP128 it's 7%.

It's actually quite significant for sorted columns for which FoR-SIMDBP128 shines (project with @krichly). Here, the NULL values can make up a much larger share.