Fix JoinIndex, add to JoinTestRunner (#1851)
This PR integrates the creation of join configurations for the JoinIndex operator and the execution of the corresponding join scenarios (closes #1755).

Compared to other join operators, the execution of JoinIndex is influenced by additional parameters (index side, indexes available). Therefore, a general function for creating more specialized configurations for operators whose executions are influenced by additional parameters is added. This function needs one basic configuration as input and returns a vector of specialized configurations as output.

In addition, the newly integrated tests for the Index Join are used to find bugs that are either fixed in this PR or documented in the issue list (closes #1852).

Bug Tracker

    no support for inequality between probe value type and indexed value type - #2077
    Semi Join, IndexSide::Left - fixed (ceb068b, c0ad173)
    incorrect results when the join mode AntiNullAsTrue is used: execution is currently not null-aware as proposed by Bellamkonda et al. [1]. - fixed (ee550e7, 9f60fb0)
    Semi/Anti Join, IndexSide::Right - fixed (6c5f0ab), removed the support (0e58af3)

Since the JoinTestRunner generates many different test cases for testing the JoinIndex, the previous test suite for the Index Join is significantly reduced.