Hash Join: Poor man's bloom filter (#2093)
In the join hash (1) create a bloom filter when materializing the build partition, (2) use that bloom filter to skip the materialization of rows on the probe side when they won't find a join partner, (3) at the same time, build a bloom filter on the build side, (4) use that second bloom filter to exclude values from the hash table that are not present on the probe side.

This is something that I built while working on the TPC-H Paper. It is not as optimized as it could be, but at least it doesn't hurt anyone.