Sort: Write Reference Tables where possible (#2085)
In most cases, it is unnecessary to materialize the result of a sort. This adds a method that returns reference tables instead of data tables.

Exception: If a column in the input references different tables or columns (e.g., as the result of a union), we still have to materialize. Also, we allow the caller to request materialization (e.g., when creating sorted tables for the benchmarks). For this, this PR fixes the materialization (fixes #2102).

The Sort test (which previously operated on tables containing only three lines) was entirely rewritten.