Stop SQLPipeline from leaking the SQLPipelineStatement(s) (#2111)
The whole idea of the SQLPipeline is to have a one stop shop for turning SQL into tables. Handing out individual pipeline statements is unnecessary. With #1993, where the transaction_context may change during the execution of a pipeline, this becomes more important as running pipeline statements individually may lead to inconsistencies.