Support Semi and Anti* joins with IndexSide::Left (#2121)
With this PR, the JoinIndex supports Semi and Anti* joins where the left input table is indexed. Even if by definition the tuples of the left input table are outputted, using indexes of the left table's chunks can be beneficial when the number of tuples of the right table is very small.