Remove ccache from CI (#2108)
Just today, I had two CI fails that were "fixed" by deleting the ccache folder and re-running the build. We originally hoped that the new ccache version fixed the issues, but it doesn't.

I'd say we retire it. This might make the CI builds a bit longer, but for me, that sounds preferrable to spuriously failing builds - either during compilation or even during execution.