Consider pruned chunk ids in LQPTranslator for index scans (#2150)
The translation of predicates node to index scans checks if a GroupKey index exists and creates the corresponding included_chunk_ids and excluded_chunk_ids vectors to specify for which chunks the index scan can be used and for which chunks the table scan should be used as a fallback.

During the creation of the include/exclude chunk id vector, information about pruned chunks are not considered at the moment. For that reason, the pull request adds an id mapping, which removes pruned chunks and adopts the chunk ids correspondingly.