Performance counters with step durations (#2073)
This PR mainly extends the operator performance counters with step runtimes, which allow storing the runtime of certain steps. The main reason has been for calibration and learning cases, but in the meanwhile I found them to be extremely helpful when analysing operator performance. The steps are shown in the tooltips of the PQP.

With this PR, one can analyze in which stages take how much time in the JoinHash. Further, and very often of interest is the trackng of segments scans on sorted segments and the number of skipped segments during execution (i.e., "dynamic" dictionary pruning, cf. Dreseler et al., PVLDB 17).

Further, this PR adds the third party module enum_magic which provides of ton of enum helper methods. An open todo is to get rid of some enum hacks (e.g., storing a count member to the get the size of enums) and use enum_magic where appropriate.
I ran TPC-H benchmarks with very small chunk counts and could not observe any performance degradations.