Remove Counting Quotient Filters (#2171)
A few months ago, I was the only one arguing to keep counting quotient filters (CQF) in Hyrise as I planned to evaluate them. As usual, I haven't found the time to do that and I don't think I'll find the time in the upcoming months.

In case they are of interest again, reverting this PR should be very simple.

The reason to remove it right now is that they don't compile on ARM servers. But that's just a side note.