Meta tables: Show allocated memory, fix line endings (#2157)
This returns the actually allocated memory. It differs from _get_process_memory_usage in that it only returns memory that has been allocated. The reported virtual memory consumption is usually higher than the amount of allocated memory as free'd memory is not immediately returned to the system (either due to internal page fragmentation or because jemalloc keeps empty pages for future use). In rare cases, it can also be higher than the amount of virtual memory if memory has been allocated but not committed yet.