Implement FDs for UnionNode (SetOperationMode::All) & Fix hashing of FDs and LQPUniqueConstraints (#2192)
This PR adds an implementation and tests for functional dependencies in UnionNode with `SetOperationMode` set to `All`.

In the course of this, various FD-related utility functions have been added as well:

* inflate_fds
* deflate_fds
* intersect_fds

---

Furthermore, the hash functions of LQPUniqueConstraint & FunctionalDependency have been fixed, so that they are no longer dependent on the internal order of elements in its unordered sets. Additional tests that cover the hash-ordering issues have been added as well.