Table Scan: Improve all-rows-match shortcut (#2194)
If the dictionary lookup shows that all values in the chunk match, we still need to make sure not to return any NULL rows. With this PR, we use the NULLable flag to skip this check where possible.

Additionally, this adds a shortcut in the generation of the output chunks if the entire chunk matches: Instead of creating a new ReferenceSegment (and potentially rebuilding the mapping), we simply forward the input chunk. This part is easier to review with whitespace changes hidden.