Hash Join: Merge output chunks <500 rows (#2202)
If a hash join's input is heavily pre-filtered or the join results in very few matches, we might end up with a high number of chunks that contain only few rows. This involves a trade-off: A lower number of output chunks reduces the overhead, especially when multi-threading is used. However, merging chunks destroys a potential references_single_chunk property of the PosList that would have been emitted otherwise.