Fix pruning of PredicateNodes with multiple outputs (#2213)
Currently, if a PredicateNode has multiple outputs, it is not considered for chunk pruning. You can see the effects in TPC-H Q4, where the predicate on the orders table is not used because the between predicate has multiple outputs at the time where the rule is run.

While we must not use predicates that are on a branch of the query, using the one on the fork is fine.

No impact on TPC-H, as the affected cases profit from being able to use a binary search and the scan is quite cheap anyway.