Sort : keep nullable property for sorted segments (#2215)
When executing ./hyriseBenchmarkTPCH -e Unencoded twice, the second run failed while loading the cached tables.
This is not due to the binary reader/writer but due to the sorting of the order and lineitem table.
The sort operator kept the nullability property for the columns but always set all segments to be nullable. This missmatch between column and semgent nullability caused the binary reader/writer to mess up.