Use const AbstractLQPNode for cardinality estimation (#2230)
AbstractCardinalityEstimator::estimate_cardinality() requires const std::shared_ptr<AbstractLQPNode>& lqp as an argument. As there is no obvious reason to make changes to an LQP while estimating the cardinality of its output, this PR changes the signature to require a shared pointer to a const LQP node and adapts the underlying data structures.