PredicateSplitUpRule: Create UnionAll for NOT LIKE (#2238)
Optimization: The ExpressionReductionRule transforms x NOT LIKE 'foo%' into x < 'foo' OR x >= 'fop'. For this special case, we know that the two OR arguments are mutually exclusive. In this case, we do not need to use SetOperationMode::Positions but can use SetOperationMode::All. This surely applies to other cases, too, but those are not covered yet.

This only applies to TPC-H Q16; other benchmarks do not use this construct.