JoinHash: Materialize smaller side first (#2239)
For the bloom filter to be efficient, it makes sense for the smaller side to be materialized first. That way, the bloom filter created on that side can serve as a filter when materializing the bigger side.

In cases where the build side is bigger than the probe side (unavoidable for some semi/anti joins), this PR flips the order of the two materialization steps.

The overall performance is not tooo impressive, as only few queries are affected. A 49% improvement for TPC-H Q22 is nice to have though.