Make hyriseBenchmarkX loading faster (#2245)
For the generation of the table histograms, use the robin_map. It allows us to cache the hash of strings, which makes generation much faster. We will likely replace the current bytell hash map with this one for other things, too.
Parallelize some steps of the benchmark start. Hide whitespaces when reviewing this.